fix logic in modal-panel.js, move large/small modal components into their own files to unbreak styleguidist
