+++ packages/wonder-blocks-modal/components/one-column-modal/one-column-modal.js
@@ -1,0 +1,42 @@
+// @flow
+import * as React from "react";
+import {MediaLayout} from "@khanacademy/wonder-blocks-layout";
+
+import LargeOneColumnModal from "./large-one-column-modal.js";
+import SmallOneColumnModal from "./small-one-column-modal.js";
+
+export type Props = {|
+    /** The modal's content. */
+    content: React.Node,
+
+    /** The optional footer to display beneath the contents. */
+    footer?: React.Node,
+
+    /**
+     * Called when the close button is clicked.
+     *
+     * If you're using `ModalLauncher`, you probably shouldn't use this prop!
+     * Instead, to listen for when the modal closes, add an `onClose` handler
+     * to the `ModalLauncher`.  Doing so will result in a console.warn().
+     */
+    onClose?: () => void,
+|};
+
+/**
+ * A one-column modal layout.
+ */
+export default class OneColumnModal extends React.Component<Props> {
+    render() {
+        return (
+            <MediaLayout>
+                {({mediaSize}) =>
+                    mediaSize === "small" ? (
+                        <SmallOneColumnModal {...this.props} />
+                    ) : (
+                        <LargeOneColumnModal {...this.props} />
+                    )
+                }
+            </MediaLayout>
+        );
+    }
+}
