+++ packages/wonder-blocks-modal/components/two-column-modal/two-column-modal.md
