+++ packages/wonder-blocks-modal/generated-snapshot.test.js
@@ -10,8 +10,8 @@ import renderer from "react-test-renderer";
 jest.mock("react-dom");
 import ModalLauncher from "./components/modal-launcher.js";
 import StandardModal from "./components/standard-modal.js";
-import TwoColumnModal from "./components/two-column-modal.js";
-import OneColumnModal from "./components/one-column-modal.js";
+import TwoColumnModal from "./components/two-column-modal/two-column-modal.js";
+import OneColumnModal from "./components/one-column-modal/one-column-modal.js";
 import ModalDialog from "./components/modal-dialog.js";
 import ModalPanel from "./components/modal-panel.js";
 import ModalContent from "./components/modal-content.js";
