+++ packages/wonder-blocks-modal/index.js
@@ -1,8 +1,8 @@
 // @flow
 import ModalLauncher from "./components/modal-launcher.js";
 import StandardModal from "./components/standard-modal.js";
-import TwoColumnModal from "./components/two-column-modal.js";
-import OneColumnModal from "./components/one-column-modal.js";
+import TwoColumnModal from "./components/two-column-modal/two-column-modal.js";
+import OneColumnModal from "./components/one-column-modal/one-column-modal.js";
 import maybeGetPortalMountedModalHostElement from "./util/maybe-get-portal-mounted-modal-host-element.js";
 
 export {
